# docs/api-reference/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pynomaly API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Pynomaly Support",
      "url": "https://pynomaly.com/support",
      "email": "support@pynomaly.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.pynomaly.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.pynomaly.com",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/detection/detect": {
      "post": {
        "tags": [
          "Anomaly Detection"
        ],
        "summary": "Detect anomalies in data",
        "description": "Detect anomalies in the provided dataset using the specified algorithm",
        "operationId": "detect_anomalies",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/DetectionRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/DetectionResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DetectionRequest": {
        "type": "object",
        "required": [
          "data",
          "algorithm"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "description": "Input data for anomaly detection",
            "example": [
              1.0,
              2.0,
              3.0,
              100.0,
              4.0,
              5.0
            ]
          },
          "algorithm": {
            "type": "string",
            "enum": [
              "isolation_forest",
              "one_class_svm",
              "lstm_autoencoder",
              "ensemble"
            ],
            "description": "Anomaly detection algorithm",
            "example": "isolation_forest"
          },
          "parameters": {
            "type": "object",
            "description": "Algorithm-specific parameters",
            "example": {
              "contamination": 0.1,
              "n_estimators": 100
            }
          }
        }
      },
      "DetectionResponse": {
        "type": "object",
        "properties": {
          "anomalies": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Indices of detected anomalies",
            "example": [
              3
            ]
          },
          "scores": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "description": "Anomaly scores for each data point",
            "example": [
              0.1,
              0.2,
              0.15,
              0.95,
              0.18,
              0.12
            ]
          },
          "threshold": {
            "type": "number",
            "description": "Anomaly detection threshold",
            "example": 0.5
          },
          "model_id": {
            "type": "string",
            "description": "ID of the model used for detection",
            "example": "isolation_forest_20240101_001"
          },
          "processing_time_ms": {
            "type": "number",
            "description": "Processing time in milliseconds",
            "example": 45.6
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Error type URI",
            "example": "https://pynomaly.com/errors/validation-error"
          },
          "title": {
            "type": "string",
            "description": "Error title",
            "example": "Validation Error"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 422
          },
          "detail": {
            "type": "string",
            "description": "Error details",
            "example": "The request data is invalid"
          },
          "instance": {
            "type": "string",
            "description": "Request instance that caused the error",
            "example": "/api/v1/detection/detect"
          }
        }
      }
    },
    "responses": {
...
      "DetectionResponse": {
        "description": "Anomaly detection response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DetectionResponse"
            }
          }
        }
      },
...
      "BadRequestError": {
        "description": "Bad request error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "type": "https://pynomaly.com/errors/bad-request",
              "title": "Bad Request",
              "status": 400,
              "detail": "Invalid request parameters"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "type": "https://pynomaly.com/errors/unauthorized",
              "title": "Unauthorized",
              "status": 401,
              "detail": "Authentication required"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "type": "https://pynomaly.com/errors/validation-error",
              "title": "Validation Error",
              "status": 422,
              "detail": "The request data is invalid"
            }
          }
        }
      },
...
    },
...
    "examples": {
      "SimpleDetectionRequest": {
        "summary": "Simple anomaly detection",
        "description": "Basic anomaly detection with Isolation Forest",
        "value": {
          "data": [
            1.0,
            2.0,
            3.0,
            100.0,
            4.0,
            5.0
          ],
          "algorithm": "isolation_forest",
          "parameters": {
            "contamination": 0.1,
            "n_estimators": 100
          }
        }
      },
      "TimeSeriesDetectionRequest": {
        "summary": "Time series anomaly detection",
        "description": "Anomaly detection for time series data",
        "value": {
          "data": [
            1.0,
            1.1,
            1.2,
            1.1,
            1.0,
            5.0,
            1.1,
            1.2
          ],
          "algorithm": "lstm_autoencoder",
          "parameters": {
            "sequence_length": 10,
            "epochs": 50
          }
        }
      },
...
    },
    "requestBodies": {
...
      "DetectionRequest": {
        "description": "Anomaly detection request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DetectionRequest"
            },
            "examples": {
              "simple": {
                "$ref": "#/components/examples/SimpleDetectionRequest"
              },
              "timeseries": {
                "$ref": "#/components/examples/TimeSeriesDetectionRequest"
              }
            }
          }
        },
        "required": true
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from the authentication endpoint"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
}